Primera version integración

+++ include/Server.hpp
@@ -27,7 +27,7 @@ class Server
 		std::string							_server_name;
 		std::string							_listen;
 		std::map<int, std::string>			_m_status_pages;
-        unsigned int						_client_max_body_size;
+        long unsigned int					_client_max_body_size;
         bool								_autoindex;
 		std::string							_index;
 		std::string							_cgi;
@@ -57,7 +57,7 @@ class Server
 		std::map<std::string, std::string> const	&getMRedirections( void ) const;
 		std::map<int, std::string> const			&getMStautsPages( void ) const;
 		std::string		 							getErrorPage( unsigned int status_code ) const;
-		unsigned int	const						&getClienteMaxBodySize( void ) const;
+		long unsigned int	const					&getClienteMaxBodySize( void ) const;
 		bool const									&getAutoindex( void )const;
         std::string const							&getIndex( void )const;
 		std::string const							&getCgi( void )const;
